# Improve allocations of the delta profiles appender