--- a/component/pyroscope/scrape/delta_profiles.go
+++ b/component/pyroscope/scrape/delta_profiles.go
@@ -4,8 +4,10 @@ import (
 	"bytes"
 	"compress/gzip"
 	"context"
+	"encoding/binary"
 	"fmt"
 	"io"
+	"sync"
 
 	"github.com/grafana/agent/component/pyroscope"
 	"github.com/grafana/agent/component/pyroscope/scrape/internal/fastdelta"
@@ -38,12 +40,15 @@ func NewDeltaAppender(appender pyroscope.Appender, labels labels.Labels) pyrosco
 		// for profiles that we don't need to produce delta, just return the appender
 		return appender
 	}
+
+	return newDeltaAppender(appender, types)
+}
+
+func newDeltaAppender(appender pyroscope.Appender, types []fastdelta.ValueType) *deltaAppender {
 	delta := &deltaAppender{
 		appender: appender,
 		delta:    fastdelta.NewDeltaComputer(types...),
-		gzw:      gzip.NewWriter(nil),
 	}
-	delta.reset()
 	return delta
 }
 
@@ -51,17 +56,59 @@ type deltaAppender struct {
 	appender pyroscope.Appender
 	delta    DeltaProfiler
 
-	buf bytes.Buffer
-	gzr gzip.Reader
-	gzw *gzip.Writer
-
 	// true if we have seen at least one sample
 	initialized bool
 }
 
-func (d *deltaAppender) reset() {
-	d.buf.Reset()
-	d.gzw.Reset(&d.buf)
+type gzipBuffer struct {
+	gzr gzip.Reader
+	gzw *gzip.Writer
+
+	out          bytes.Buffer
+	uncompressed bytes.Buffer
+	in           *bytes.Reader
+}
+
+var gzipBufferPool = sync.Pool{
+	New: func() interface{} {
+		return &gzipBuffer{
+			gzw: gzip.NewWriter(nil),
+			in:  bytes.NewReader(nil),
+		}
+	},
+}
+
+func getGzipBuffer() *gzipBuffer {
+	buf := gzipBufferPool.Get().(*gzipBuffer)
+	buf.reset()
+	return buf
+}
+
+func putGzipBuffer(buf *gzipBuffer) {
+	gzipBufferPool.Put(buf)
+}
+
+func (d *gzipBuffer) reset() io.Writer {
+	d.out.Reset()
+	d.gzw.Reset(&d.out)
+	return d.gzw
+}
+
+func (d *gzipBuffer) uncompress(in []byte) ([]byte, error) {
+	if !isGzipData(in) {
+		return in, nil
+	}
+	d.in.Reset(in)
+	if err := d.gzr.Reset(d.in); err != nil {
+		return nil, err
+	}
+	d.uncompressed.Reset()
+	d.uncompressed.Grow(uncompressedSize(in))
+	_, err := d.uncompressed.ReadFrom(&d.gzr)
+	if err != nil {
+		return nil, fmt.Errorf("decompressing profile: %v", err)
+	}
+	return d.uncompressed.Bytes(), nil
 }
 
 func (d *deltaAppender) Append(ctx context.Context, lbs labels.Labels, samples []*pyroscope.RawSample) error {
@@ -89,32 +136,36 @@ func (d *deltaAppender) Append(ctx context.Context, lbs labels.Labels, samples [
 // data is uncompressed if it is gzip compressed.
 // The returned data is always gzip compressed.
 func (d *deltaAppender) computeDelta(data []byte) (b []byte, err error) {
-	if isGzipData(data) {
-		if err := d.gzr.Reset(bytes.NewReader(data)); err != nil {
-			return nil, err
-		}
-		data, err = io.ReadAll(&d.gzr)
-		if err != nil {
-			return nil, fmt.Errorf("decompressing profile: %v", err)
-		}
+	gzipBuf := getGzipBuffer()
+	defer putGzipBuffer(gzipBuf)
+
+	data, err = gzipBuf.uncompress(data)
+	if err != nil {
+		return nil, err
 	}
 
-	d.reset()
-
-	if err = d.delta.Delta(data, d.gzw); err != nil {
+	if err = d.delta.Delta(data, gzipBuf.gzw); err != nil {
 		return nil, fmt.Errorf("computing delta: %v", err)
 	}
-	if err := d.gzw.Close(); err != nil {
+	if err := gzipBuf.gzw.Close(); err != nil {
 		return nil, fmt.Errorf("closing gzip writer: %v", err)
 	}
 	// The returned slice will be retained in case the profile upload fails,
 	// so we need to return a copy of the buffer's bytes to avoid a data
 	// race.
-	b = make([]byte, len(d.buf.Bytes()))
-	copy(b, d.buf.Bytes())
+	b = make([]byte, len(gzipBuf.out.Bytes()))
+	copy(b, gzipBuf.out.Bytes())
 	return b, nil
 }
 
 func isGzipData(data []byte) bool {
 	return bytes.HasPrefix(data, []byte{0x1f, 0x8b})
+}
+
+func uncompressedSize(in []byte) int {
+	last := len(in)
+	if last < 4 {
+		return -1
+	}
+	return int(binary.LittleEndian.Uint32(in[last-4 : last]))
 }