This improves allocations of the delta profiles appender.

Mostly it doesn't use `ioutil.ReadAll` but instead grows and reuses a buffer. I've also moved the buffer to a pool to avoid having those scale with the amount of targets.

Result on a very small profile.

```
❯ benchstat before.txt after.txt
name             old time/op    new time/op    delta
ComputeDelta-16    31.1µs ± 5%    31.1µs ± 3%     ~     (p=0.842 n=9+10)

name             old alloc/op   new alloc/op   delta
ComputeDelta-16      743B ± 0%      169B ± 8%  -77.22%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
ComputeDelta-16      3.00 ± 0%      1.00 ± 0%  -66.67%  (p=0.000 n=10+10)
```